backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
mediapipe
pyaudio
pyfluidsynth
SpeechRecognition
pymongo
dnspython
motor
python-dotenv
pydantic
pydantic-settings
email-validator
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
scipy
numpy
